Use new loki API
Switch to the new loki API As the old prom API is
[deprecated](https://github.com/grafana/loki/blob/master/docs/api.md#get-apipromquery).
This uses the client from the logcli package so we don't need to worry
about marshaling/unmarshaling anymore. Regexp filtering via parameter is
not possible anymore using the new API so that field has been removed.
Filtering should be done directly in the query with logql.